PostgreSQL Faz 2 Adim 4 (cekirdek): 52 migration Neon PostgreSQL'de temiz kuruluyor
GERCEK SINAV GECILDI: tum semanin PostgreSQL'de kuruldugu dogrulandi.
PostgresMigrationTests bos Neon DB'sinde 52 migration'i calistirir ve dogrular.

Cikan ve cozulen PostgreSQL farklari (hepsi SQLite'ta da calisir):
1. 22003 integer out of range: SQLite INTEGER 64-bit, PG 32-bit. Unix-ms zaman
   damgalari (~1.7 trilyon) tasiyordu -> tum sema INTEGER -> BIGINT (SQLite'ta
   BIGINT = INTEGER affinity, davranis ayni).
2. 42601 PRAGMA / 42883 pragma_table_info: SQLite sema-sorgulama PostgreSQL'de
   yok. Yeni DbIntrospect yardimcisi (lehce-duyarli): ColumnExists/TableExists/
   ColumnNames/PrimaryKey/ListTables SQLite PRAGMA <-> PG information_schema.
   8 migration + BusinessSyncService + CompanyPurgeService buna baglandi.
3. SQLite CreateCommand() acik transaction'i OTOMATIK atar; DbIntrospect null tx
   ile ezip 'pending transaction' hatasi veriyordu -> yalniz gercek tx verilince ata.

DOGRULAMA: 572 test yesil (569 SQLite + 2 Neon baglanti + 1 migration), 0 hata,
0 atlanan. Masaustu (SQLite) davranisi degismedi.

Not: DatabaseHealth (PRAGMA journal/foreign_keys) masaustu-SQLite'a ozel saglik
kontrolu; sunucu-PG wiring'inde (Faz 3) ele alinacak, migration'i etkilemez.

Siradaki: uctan uca PostgreSQL veri islemleri (servisler PG'de calisiyor mu).

+++ src/DepoWise.Infrastructure/Database/Migrations/Migration002_AuthSeed.cs
@@ -19,8 +19,8 @@ CREATE TABLE login_attempts (
     id TEXT PRIMARY KEY,
     company_id TEXT NULL,
     username TEXT NOT NULL,
-    success INTEGER NOT NULL,
-    attempted_at INTEGER NOT NULL
+    success BIGINT NOT NULL,
+    attempted_at BIGINT NOT NULL
 );
 CREATE INDEX ix_login_attempts_user ON login_attempts(username, attempted_at);
 
@@ -28,9 +28,9 @@ CREATE TABLE sessions (
     id TEXT PRIMARY KEY,
     user_id TEXT NOT NULL,
     company_id TEXT NOT NULL,
-    created_at INTEGER NOT NULL,
-    expires_at INTEGER NOT NULL,
-    revoked_at INTEGER NULL,
+    created_at BIGINT NOT NULL,
+    expires_at BIGINT NOT NULL,
+    revoked_at BIGINT NULL,
     FOREIGN KEY (user_id) REFERENCES users(id)
 );
 CREATE INDEX ix_sessions_user ON sessions(user_id);
