+++ src/DepoWise.Infrastructure/Database/Migrations/Migration013_RememberTokens.cs
@@ -21,8 +21,8 @@ CREATE TABLE remember_tokens (
     user_id TEXT NOT NULL,
     company_id TEXT NOT NULL,
     token_hash TEXT NOT NULL,
-    expires_at INTEGER NOT NULL,
-    created_at INTEGER NOT NULL,
+    expires_at BIGINT NOT NULL,
+    created_at BIGINT NOT NULL,
     FOREIGN KEY (user_id) REFERENCES users(id)
 );
 CREATE INDEX ix_remember_tokens ON remember_tokens(token_hash, expires_at);";
